test: decouple the stern multitailer
Multitailer now isn't tied to the environment it is running into,
that knowledge is now in the environment.

Also, make the multitailer safe for multiple calls.
