+++ tests/utils/sternmultitailer/multitailer.go
@@ -17,55 +17,181 @@ limitations under the License.
 package sternmultitailer
 
 import (
+	"bufio"
 	"context"
+	"encoding/json"
+	"fmt"
 	"io"
 	"os"
+	"path"
+	"regexp"
+	"text/template"
+	"time"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"github.com/stern/stern/stern"
+	"k8s.io/apimachinery/pkg/fields"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/kubernetes"
-
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 )
 
-const (
-	// ClusterLogsDirectory contains the fixed path to store the cluster logs
-	ClusterLogsDirectory = "cluster_logs/"
+// SternMultiTailer can tail logs from multiple pods at once using stern, and write them to disk
+// in a structured way.
+type SternMultiTailer struct{}
 
-	// OperatorLogsDirectory is the fixed path to store the logs of all the operator pods
-	OperatorLogsDirectory = "operator_logs/"
-)
+// StreamLogs opens a goroutine to execute stern on all the pods that match
+// the labelSelector. Their logs will be written to disk in the outputBaseDir, split by namespace, pod and container,
+// using a namespace/pod/container.log file for each container of matching pods.
+// Close the ctx context to terminate stern execution.
+// Returns a channel that will be closed when all the logs have been written to disk
+// and the ones we asked to remove have been deleted.
+func (s *SternMultiTailer) StreamLogs(
+	ctx context.Context,
+	client kubernetes.Interface,
+	labelSelector labels.Selector,
+	outputBaseDir string,
+) chan struct{} {
+	outPipeReader, outPipeWriter := io.Pipe()
+	errOut := os.Stdout
 
-// SternMultiTailer contains the necessary data for the logs of every cluster
-type SternMultiTailer struct {
-	stdOut       *io.PipeReader
-	openFilesMap map[string]*os.File
+	// JSON output
+	pod := regexp.MustCompile(".*")
+	container := regexp.MustCompile(".*")
+	t := "{ \"message\": {{json .Message}}, " +
+		"\"namespace\": \"{{.Namespace}}\", " +
+		"\"podName\": \"{{.PodName}}\", " +
+		"\"containerName\": \"{{.ContainerName}}\" }\n"
+
+	funs := template.FuncMap{
+		"json": func(v interface{}) (string, error) {
+			b, err := json.Marshal(v)
+			if err != nil {
+				return "", err
+			}
+			return string(b), nil
+		},
+	}
+
+	parsedTemplate, _ := template.New("log").Funcs(funs).Parse(t)
+
+	config := &stern.Config{
+		Namespaces:            []string{},
+		PodQuery:              pod,
+		ContainerQuery:        container,
+		ExcludePodQuery:       []*regexp.Regexp{},
+		Timestamps:            false,
+		TimestampFormat:       stern.TimestampFormatDefault,
+		Location:              time.UTC,
+		ExcludeContainerQuery: []*regexp.Regexp{},
+		ContainerStates: []stern.ContainerState{
+			stern.RUNNING,
+		},
+		Exclude:             []*regexp.Regexp{},
+		Include:             []*regexp.Regexp{},
+		Highlight:           []*regexp.Regexp{},
+		InitContainers:      true,
+		EphemeralContainers: true,
+		Since:               48 * time.Hour,
+		AllNamespaces:       true,
+		LabelSelector:       labelSelector,
+		FieldSelector:       fields.Everything(),
+		TailLines:           nil,
+		Template:            parsedTemplate,
+		Follow:              true,
+		Resource:            "",
+		OnlyLogLines:        true,
+		MaxLogRequests:      50,
+		Stdin:               false,
+		DiffContainer:       false,
+
+		Out:    outPipeWriter,
+		ErrOut: errOut,
+	}
+
+	outputDone := make(chan struct{})
+	go func() {
+		err := stern.Run(ctx, client, config)
+		if err != nil {
+			fmt.Printf("stern failed: %v", err)
+		}
+		if <-ctx.Done(); true {
+			_ = outPipeWriter.Close()
+			_ = errOut.Close()
+		}
+	}()
+
+	go func() {
+		s.outputWriter(outputBaseDir, outPipeReader)
+		close(outputDone)
+	}()
+
+	return outputDone
 }
 
-// CatchClusterLogs execute StreamLogs with the specific labels to match
-// only the CNPG pods and send them to ClusterLogsDirectory path
-func (s *SternMultiTailer) CatchClusterLogs(ctx context.Context, client kubernetes.Interface) chan struct{} {
-	// Select all the pods belonging to CNPG
-	labelSelector, _ := metav1.LabelSelectorAsSelector(&metav1.LabelSelector{
-		MatchExpressions: []metav1.LabelSelectorRequirement{
-			{
-				Key:      utils.ClusterLabelName,
-				Operator: metav1.LabelSelectorOpExists,
-			},
-		},
-	})
+func (s *SternMultiTailer) outputWriter(baseDir string, logReader io.Reader) {
+	r := bufio.NewReader(logReader)
+	openFilesMap := make(map[string]*os.File)
+	defer func() {
+		for k, file := range openFilesMap {
+			_ = file.Close()
+			delete(openFilesMap, k)
+		}
+	}()
+	for {
+		lineBytes, readErr := r.ReadBytes('\n')
+		// If we have a read error, skip the line
+		if readErr != nil && readErr != io.EOF {
+			fmt.Printf("could not read log line from pipe: %v\n", readErr)
+			continue
+		}
 
-	return s.streamLogs(ctx, client, labelSelector, ClusterLogsDirectory)
+		// If we have an EOF and the line is empty, I'm done
+		if readErr == io.EOF && len(lineBytes) == 0 {
+			break
+		}
+
+		// Otherwise, we have a line to process
+		var logLine stern.Log
+		err := json.Unmarshal(lineBytes, &logLine)
+		if err != nil {
+			fmt.Printf("could not unmarshal log line %v: %v\n", logLine, err)
+			continue
+		}
+
+		file, err := s.getLogFile(baseDir, logLine, openFilesMap)
+		if err != nil {
+			fmt.Printf("no file to write log line %v: %v\n", logLine, err)
+			continue
+		}
+
+		_, err = file.WriteString(fmt.Sprintf("%v\n", logLine.Message))
+		if err != nil {
+			fmt.Printf("could not write message to file %v: %v\n", file.Name(), err)
+			continue
+		}
+	}
 }
 
-// CatchOperatorLogs execute streamLogs with the labels to match the Operator labels
-// and send them to OperatorLogsDirectory
-func (s *SternMultiTailer) CatchOperatorLogs(ctx context.Context, client kubernetes.Interface) chan struct{} {
-	// Select all the pods belonging to CNPG
-	labelSelector, _ := metav1.LabelSelectorAsSelector(&metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"app.kubernetes.io/name": "cloudnative-pg",
-		},
-	})
+// Get an open file for the log, or open a new one
+func (s *SternMultiTailer) getLogFile(baseDir string, log stern.Log, openFilesMap map[string]*os.File) (*os.File,
+	error) {
+	filePath := path.Join(baseDir, log.Namespace, log.PodName, log.ContainerName+".log")
+	dirFile := path.Dir(filePath)
 
-	return s.streamLogs(ctx, client, labelSelector, OperatorLogsDirectory)
+	file, ok := openFilesMap[filePath]
+	if ok {
+		return file, nil
+	}
+
+	// If we don't have the file already opened, we open it
+	err := os.MkdirAll(dirFile, 0o700)
+	if err != nil {
+		return nil, fmt.Errorf("cannot ensure directory existence (%v): %w", dirFile, err)
+	}
+	file, err = os.OpenFile(filePath, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0o600) // nolint:gosec
+	if err != nil {
+		return nil, fmt.Errorf("cannot open file %v: %w", filePath, err)
+	}
+
+	openFilesMap[filePath] = file
+	return file, nil
 }
